P4ADEV-2585 fix pr
