+++ src/test/java/it/gov/pagopa/pu/pagopapayments/mapper/NoticeRequestMapperTest.java
@@ -3,7 +3,6 @@ package it.gov.pagopa.pu.pagopapayments.mapper;
 import it.gov.pagopa.pu.debtpositions.dto.generated.InstallmentDTO;
 import it.gov.pagopa.pu.debtpositions.dto.generated.PersonDTO;
 import it.gov.pagopa.pu.pagopapayments.util.TestUtils;
-import it.gov.pagopa.pu.printpaymentnotice.connector.printpaymentnotice.generated.dto.DebtorDTO;
 import it.gov.pagopa.pu.printpaymentnotice.connector.printpaymentnotice.generated.dto.NoticeDTO;
 import it.gov.pagopa.pu.printpaymentnotice.connector.printpaymentnotice.generated.dto.NoticeRequestDataDTO;
 import org.junit.jupiter.api.Assertions;
@@ -18,74 +17,6 @@ class NoticeRequestMapperTest {
 
   public NoticeRequestMapperTest() {
     this.podamFactory = TestUtils.getPodamFactory();
-  }
-
-  @Test
-  void givenValidPersonWhenToDebtorDTOThenOk() {
-    //given
-    PersonDTO personRequest = podamFactory.manufacturePojo(PersonDTO.class);
-
-    //when
-    DebtorDTO response = NoticeRequestMapper.toDebtorDTO(personRequest);
-
-    //verify
-    assertNotNull(response);
-    Assertions.assertEquals(personRequest.getAddress(), response.getAddress());
-    Assertions.assertEquals(personRequest.getCivic(), response.getBuildingNumber());
-    Assertions.assertEquals(personRequest.getLocation(), response.getCity());
-    Assertions.assertEquals(personRequest.getFullName(), response.getFullName());
-    Assertions.assertEquals(personRequest.getPostalCode(), response.getPostalCode());
-    Assertions.assertEquals(personRequest.getProvince(), response.getProvince());
-  }
-
-  @Test
-  void givenInvalidAddressWhenToDebtorDTOThenError() {
-    //given
-    PersonDTO personRequest = podamFactory.manufacturePojo(PersonDTO.class);
-    personRequest.setAddress(null);
-
-    //when & verify
-    Assertions.assertThrows(IllegalArgumentException.class, () -> NoticeRequestMapper.toDebtorDTO(personRequest));
-  }
-
-  @Test
-  void givenInvalidCivicWhenToDebtorDTOThenError() {
-    //given
-    PersonDTO personRequest = podamFactory.manufacturePojo(PersonDTO.class);
-    personRequest.setCivic(null);
-
-    //when & verify
-    Assertions.assertThrows(IllegalArgumentException.class, () -> NoticeRequestMapper.toDebtorDTO(personRequest));
-  }
-
-  @Test
-  void givenInvalidLocationWhenToDebtorDTOThenError() {
-    //given
-    PersonDTO personRequest = podamFactory.manufacturePojo(PersonDTO.class);
-    personRequest.setLocation(null);
-
-    //when & verify
-    Assertions.assertThrows(IllegalArgumentException.class, () -> NoticeRequestMapper.toDebtorDTO(personRequest));
-  }
-
-  @Test
-  void givenInvalidPostalCodeWhenToDebtorDTOThenError() {
-    //given
-    PersonDTO personRequest = podamFactory.manufacturePojo(PersonDTO.class);
-    personRequest.setPostalCode(null);
-
-    //when & verify
-    Assertions.assertThrows(IllegalArgumentException.class, () -> NoticeRequestMapper.toDebtorDTO(personRequest));
-  }
-
-  @Test
-  void givenInvalidProvinceWhenToDebtorDTOThenError() {
-    //given
-    PersonDTO personRequest = podamFactory.manufacturePojo(PersonDTO.class);
-    personRequest.setProvince(null);
-
-    //when & verify
-    Assertions.assertThrows(IllegalArgumentException.class, () -> NoticeRequestMapper.toDebtorDTO(personRequest));
   }
 
   @Test
@@ -134,7 +65,7 @@ class NoticeRequestMapperTest {
     //verify
     assertNotNull(response);
     Assertions.assertEquals(NoticeRequestMapper.toNoticeDTO(installmentRequest), response.getNotice());
-    Assertions.assertEquals(NoticeRequestMapper.toDebtorDTO(personRequest), response.getDebtor());
+    Assertions.assertEquals(DebtorMapper.toDebtorDTO(personRequest), response.getDebtor());
     Assertions.assertEquals(taxCode, response.getCreditorInstitution().getTaxCode());
   }
 }
